pyproject.toml
[project]
name = "neurapose-app"
version = "1.0.0"
description = "NeuraPose é um sistema de visão computacional que utiliza técnicas de aprendizagem profunda para a detecção de posturas humana usando modelos temporais como LSTM e modelos de detecção de objetos como YOLO."
authors = [
    { name = "Cecilius Afonso Sousa do Carmo", email = "ceciliusousa@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.10.11"
dependencies = [
    "aiofiles>=25.1.0",
    "colorama==0.4.6",
    "fastapi>=0.128.0",
    "gdown==5.2.0",
    "matplotlib==3.10.8",
    "onnxruntime-gpu==1.23.2",
    "opencv-python==4.12.0.88",
    "pandas==2.3.3",
    "polars==1.36.1",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "scikit-learn==1.7.2",
    "seaborn==0.13.2",
    "tensorboard==2.20.0",
    "torch>=2.9.1",
    "torchreid==0.2.5",
    "torchvision>=0.24.1",
    "tqdm==4.67.1",
    "ultralytics==8.3.240",
    "uvicorn[standard]>=0.40.0",
    "yt-dlp==2025.12.8",
]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[[tool.uv.index]]
url = "https://pypi.org/simple"
default = true
